Remove re-retrieve of window
Re-retrieving the window was added to be able to handle some
unreachable elements. However, the real issue seems to lie with the
Windows UIA itself, and the workaround is to retrieve elements from
their parent step by step instead of from their ancestor or from their
window parent. Therefore removing this code, that also relied on making
the property `IsMainWindow` public in FlaUI.