Reinstate usage of GetAllTopLevelWindows.
This was removed due to https://github.com/FlaUI/FlaUI/issues/596, which I would like to investigate, however I've so far been unable to repro the issue. Maybe if the failing code path is enabled then I'll hit it and will be able to investigate.

Also, I need to be able to retrieve all open window handles, so only returning the main window and modal windows is a problem for me...