Add input sources in GetOrCreateInputSource
The spec is unclear on when input sources should be added to the input source map. Previously it was being done in `ExtractActionSequence` but the problem with doing it there is that there's no way to know if `GetOrCreateInputSource` is returning an existing input source or creating a new one, causing #43.

Move the logic to add an input source to `GetOrCreateInputSource` to fix  this.

(It would kinda make sense to do it instead in `CreateInputSource` but the spec defines a caller of this method which adds the input source manually :/ )

Fixes #43