Ignore not matched capabilities
Instead of returning 'session not created' when not all capabilities
were matched, ignore not matching capabilities. I am not sure whether
this is to the letter of the standard, but it will not do much harm
except for users that may be surprised if FlaUI.WebDriver just ignores
their supplied capability which is not implemented for FlaUI.WebDriver
yet.

Only unknown extension capabilities are ignored.

Fixes #100.